Make cluster itself iterable, instead of managing cycles via a class. Update tests.
